created prepareDataFactory; refactored prepareData for objects
