+++ frontend/src/views/assignments/import-export.js
@@ -12,185 +12,16 @@ import { useSelector, useDispatch } from "react-redux";
 import { ExportActionButton } from "../../components/export-button";
 import { ImportActionButton } from "../../components/import-button";
 import { Alert } from "react-bootstrap";
-import { normalizeImport, dataToFile } from "../../libs/importExportUtils";
+import {
+    normalizeImport,
+    prepareDataFactory,
+} from "../../libs/importExportUtils";
 import {
     AssignmentsList,
     AssignmentsDiffList,
 } from "../../components/assignments-list";
-import { prepareMinimal } from "../../libs/exportUtils";
 import { diffImport, getChanged } from "../../libs/diffUtils";
 import { offerTableSelector } from "../offertable/actions";
-
-/**
- * Format a date as YYYY-MM-DD for inserting into a spreadsheet
- *
- * @param {*} date
- * @returns
- */
-function formatDateForSpreadsheet(date) {
-    try {
-        return date && new Date(date).toJSON().slice(0, 10);
-    } catch (e) {
-        return "";
-    }
-}
-
-/**
- * Create header columns for a spreadsheet containing information about every pay period.
- *
- * @param {*} assignments
- * @returns
- */
-function createPayPeriodHeaders(assignments) {
-    const ret = [];
-    if (!assignments) {
-        return ret;
-    }
-    const maxNumPeriods = Math.max(
-        ...assignments.map((assignment) => assignment.wage_chunks?.length || 0),
-        0
-    );
-
-    for (let i = 0; i < maxNumPeriods; i++) {
-        ret.push(
-            `Period ${i + 1} Rate`,
-            `Period ${i + 1} Hours`,
-            `Period ${i + 1} Start Date`,
-            `Period ${i + 1} End Date`
-        );
-    }
-    return ret;
-}
-
-/**
- * Create formatted rows providing information about each wage chunk.
- *
- * @param {*} wageChunks
- * @returns
- */
-function formatWageChunksToList(wageChunks) {
-    const ret = [];
-    if (!wageChunks) {
-        return ret;
-    }
-
-    ret.push(wageChunks.length);
-    for (const chunk of wageChunks) {
-        ret.push(
-            chunk.rate,
-            chunk.hours,
-            formatDateForSpreadsheet(chunk.start_date),
-            formatDateForSpreadsheet(chunk.end_date)
-        );
-    }
-    return ret;
-}
-
-/**
- * Returns a function which converts a list of selected assignments into a File object
- *
- * @export
- * @returns
- */
-export function prepareData(
-    selectedAssignmentIds,
-    session,
-    createPayPeriodHeaders,
-    formatWageChunksToList,
-    formatDateForSpreadsheet
-) {
-    /**
-     * Converts a list of assignments into a File object
-     *
-     * @export
-     * @returns
-     */
-    function prepareSelectedData(assignments, dataFormat) {
-        // If we have selected specific assignments, we only want to export those.
-        if (selectedAssignmentIds && selectedAssignmentIds.length > 0) {
-            assignments = assignments.filter((a) =>
-                selectedAssignmentIds.includes(a.id)
-            );
-        }
-
-        // We want to flatten a lot of the data in `assignments` and only include the information
-        // we need.
-        const assignmentsForSpreadsheet = assignments.map((assignment) => ({
-            first_name: assignment.applicant.first_name,
-            last_name: assignment.applicant.last_name,
-            utorid: assignment.applicant.utorid,
-            email: assignment.applicant.email,
-            position_code: assignment.position.position_code,
-            start_date: assignment.start_date,
-            end_date: assignment.end_date,
-            contract_template: assignment.contract_override_pdf
-                ? null
-                : assignment.position.contract_template.template_name,
-            contract_override_pdf: assignment.contract_override_pdf,
-            hours: assignment.hours,
-            active_offer_status: assignment.active_offer_status,
-            active_offer_recent_activity_date:
-                assignment.active_offer_recent_activity_date,
-            wage_chunks: assignment.wage_chunks.map((chunk) => ({
-                hours: chunk.hours,
-                rate: chunk.rate,
-                start_date: chunk.start_date,
-                end_date: chunk.end_date,
-            })),
-        }));
-        return dataToFile(
-            {
-                toSpreadsheet: () =>
-                    [
-                        [
-                            "Last Name",
-                            "First Name",
-                            "UTORid",
-                            "Email",
-                            "Position Code",
-                            "Start Date",
-                            "End Date",
-                            "Hours",
-                            "Contract Template",
-                            "Contract Override PDF",
-                            "Offer Status",
-                            "Recent Activity Date",
-                            "",
-                            "Number of Pay Periods",
-                            ...createPayPeriodHeaders(
-                                assignmentsForSpreadsheet
-                            ),
-                        ],
-                    ].concat(
-                        assignmentsForSpreadsheet.map((assignment) => [
-                            assignment.last_name,
-                            assignment.first_name,
-                            assignment.utorid,
-                            assignment.email,
-                            assignment.position_code,
-                            formatDateForSpreadsheet(assignment.start_date),
-                            formatDateForSpreadsheet(assignment.end_date),
-                            assignment.hours,
-                            assignment.contract_template,
-                            assignment.contract_override_pdf,
-                            assignment.active_offer_status,
-                            undefined,
-                            assignment.active_offer_recent_activity_date,
-                            ...formatWageChunksToList(assignment.wage_chunks),
-                        ])
-                    ),
-                toJson: () => ({
-                    assignments: assignments.map((assignment) =>
-                        prepareMinimal.assignment(assignment, session)
-                    ),
-                }),
-            },
-            dataFormat,
-            "assignments"
-        );
-    }
-    return prepareSelectedData;
-}
 
 /**
  * Allows for the download of a file blob containing the exported instructors.
@@ -231,12 +62,18 @@ export function ConnectedExportAssignmentsAction({
             setInProgress(true);
             const file = await dispatch(
                 exportAssignments(
-                    prepareData(
-                        selectedAssignmentIds,
-                        session,
-                        createPayPeriodHeaders,
-                        formatWageChunksToList,
-                        formatDateForSpreadsheet
+                    prepareDataFactory(
+                        "assignment",
+                        (assignments, selectedIds = selectedAssignmentIds) => {
+                            // If we have selected specific assignments, we only want to export those.
+                            if (selectedIds && selectedIds.length > 0) {
+                                assignments = assignments.filter((a) =>
+                                    selectedIds.includes(a.id)
+                                );
+                            }
+                            return assignments;
+                        },
+                        session
                     ),
                     exportType
                 )
